+++ src/modules/flat/message.ts
@@ -1,10 +1,11 @@
 import type { ChatMessagePF2e, SpellPF2e } from "foundry-pf2e"
 import { MODULE_ID } from "src/constants"
 import MODULE from "src/index"
-import { translate } from "src/utils"
+import { parseHTML, translate } from "src/utils"
 import { BaseModule } from "../base"
-import { flatMessageConfig } from "./message-config"
-import { flatCheckForUserTargets } from "./target"
+import { collectFlatChecks, type FlatCheckData } from "./data"
+import { localizeCondition, localizeSource } from "./i18n"
+import { setupRuleElements } from "./rules/setup"
 
 export class MessageFlatCheckModule extends BaseModule {
 	settingsKey = "flat-check-in-message"
@@ -16,62 +17,37 @@ export class MessageFlatCheckModule extends BaseModule {
 		this.registerHook("preCreateChatMessage", preCreateMessage)
 		this.registerWrapper("ChatMessage.prototype.renderHTML", messageGetHTMLWrapper, "WRAPPER")
 		this.registerSocket("flat-dsn", handleDSNSocket)
+
+		setupRuleElements()
 	}
 	disable() {
 		super.disable()
 	}
-	onReady() {
-		setTimeout(() => {
-			if (
-				game.modules.get("pf2e-perception")?.active &&
-				!(
-					(game.modules.get("pf2e-perception") as any)?.api?.check?.getFlatCheckDc instanceof
-					Function
-				)
-			) {
-				foundry.applications.api.DialogV2.prompt({
-					window: { title: translate("flat.message.perception-outdated-title") },
-					content: translate("flat.message.perception-outdated-content"),
-					ok: {
-						label: translate("flat.message.button-close"),
-						icon: "fas fa-close",
-					},
-				})
-			}
-		}, 1000)
-	}
 }
 
-const REROLL_ICONS = {
+type RerollModes = "low" | "high" | "hero" | "new"
+const REROLL_ICONS: Record<RerollModes, string> = {
 	hero: "fa-solid fa-hospital-symbol",
 	new: "fa-solid fa-dice",
 	low: "fa-solid fa-dice-one",
-	higher: "fa-solid fa-dice-six",
+	high: "fa-solid fa-dice-six",
 }
 
-interface ButtonData {
-	label: string
-	description?: string
-	dc: number
+type MsgFlagDataEntry = FlatCheckData & {
 	roll?: number
 	reroll?: {
 		oldRoll: number
-		keep: "low" | "high" | "hero" | "new"
+		keep: RerollModes
 	}
 }
 
-interface ButtonsFlags {
-	grabbed?: ButtonData
-	stupefied?: ButtonData
-	deafened?: ButtonData
-	targets?: ButtonData | { count: number }
-}
+type MsgFlagData = Record<string, MsgFlagDataEntry>
 
 export async function messageGetHTMLWrapper(this: ChatMessagePF2e, wrapper, ...args) {
 	const html: HTMLElement = await wrapper(...args)
 
 	try {
-		if (this.isContentVisible) renderButtons(this, html)
+		if (this.isContentVisible) await renderButtons(this, html)
 	} catch (e) {
 		console.error("Exception occured while rendering message flat-check buttons: ", e)
 	}
@@ -79,83 +55,86 @@ export async function messageGetHTMLWrapper(this: ChatMessagePF2e, wrapper, ...a
 	return html
 }
 
-function renderButtons(msg: ChatMessagePF2e, html: HTMLElement) {
-	const buttons: Array<(ButtonData & { key: string }) | { text: string }> = []
-	const data = msg.flags[MODULE_ID]?.flatchecks as ButtonsFlags | undefined
-	if (data) {
-		if (data.grabbed) buttons.push({ key: "grabbed", ...data.grabbed })
-		if (data.stupefied) buttons.push({ key: "stupefied", ...data.stupefied })
-		if (data.deafened) buttons.push({ key: "deafened", ...data.deafened })
-		if (data.targets) {
-			if ("count" in data.targets)
-				buttons.push({
-					text: translate("flat.message.button-require-flat-check", { count: data.targets.count }),
-				})
-			else buttons.push({ key: "targets", ...data.targets })
+async function renderButtons(msg: ChatMessagePF2e, html: HTMLElement) {
+	interface ButtonData {
+		key: string
+		condition: string
+		source: string
+		baseDc: number
+		finalDc: number
+		adjustments?: string
+		rolls: { class: string; value: number }[]
+		rerollIcon?: string
+		showRollButton: boolean
+	}
+
+	const data = msg.flags[MODULE_ID]?.flatchecks as MsgFlagData | undefined
+	if (!data) return
+
+	const buttons: ButtonData[] = []
+
+	for (const [key, check] of Object.entries(data)) {
+		const rollData: { class: string; value: number }[] = []
+		const rolls: [number | undefined, number | undefined] = [check.reroll?.oldRoll, check.roll]
+
+		const rollClasses: [string, string] = ["strikethrough", "strikethrough"]
+		if (check.reroll?.keep === "high") {
+			const higherIndex = check.reroll.oldRoll < check.roll! ? 1 : 0
+			const outcome = rolls[higherIndex]! >= check.finalDc ? "success" : "failure"
+			rollClasses[higherIndex] = outcome
+		} else if (check.reroll?.keep === "low") {
+			const lowerIndex = check.reroll.oldRoll > check.roll! ? 1 : 0
+			const outcome = rolls[lowerIndex]! >= check.finalDc ? "success" : "failure"
+			rollClasses[lowerIndex] = outcome
+		} else {
+			const outcome = check.roll ? (check.roll >= check.finalDc ? "success" : "failure") : ""
+			rollClasses[1] = outcome
 		}
-	} else {
-		return
+
+		if (check.reroll?.oldRoll) {
+			rollData.push({ class: rollClasses[0], value: check.reroll.oldRoll })
+		}
+		if (check.roll) {
+			rollData.push({ class: rollClasses[1], value: check.roll })
+		}
+
+		buttons.push({
+			baseDc: check.baseDc,
+			finalDc: check.finalDc,
+			adjustments: check.adjustments?.map((a) => `${a.label} ${a.value}`).join("<br>"),
+			condition: check.condition,
+			source: check.source,
+			rolls: rollData,
+			rerollIcon: check.reroll?.keep ? REROLL_ICONS[check.reroll?.keep] : undefined,
+			showRollButton:
+				msg.canUserModify(game.user as unknown as foundry.documents.BaseUser, "update") &&
+				!check.reroll,
+			key,
+		})
 	}
 
 	if (buttons.length) {
-		const buttonHtml = buttons.map((data) => {
-			if ("text" in data) {
-				return `<div class="fc-note"><i class="fa-regular fa-circle-question"></i> ${data.text}</div>`
-			}
+		const renderData = {
+			buttons,
+			i18n: (key: string) => {
+				return translate(`flat.${key}`)
+			},
+			localizeCondition,
+			localizeSource,
+		}
 
-			const buttonIcon = data.roll ? "fa-rotate rotate" : "fa-dice-d20 die"
-			const buttonClass =
-				msg.canUserModify(game.user as unknown as foundry.documents.BaseUser, "update") &&
-				!data.reroll
-					? ""
-					: "hidden"
+		const buttonHtml = await foundry.applications.handlebars.renderTemplate(
+			"modules/pf2e-flatcheck-helper/templates/flat-check-buttons.hbs",
+			renderData,
+		)
 
-			const rolls: [number | undefined, number | undefined] = [data.reroll?.oldRoll, data.roll]
-			const rollClasses: [string, string] = ["strikethrough", "strikethrough"]
-			if (data.reroll?.keep === "high") {
-				const higherIndex = data.reroll.oldRoll < data.roll! ? 1 : 0
-				const outcome = rolls[higherIndex]! >= data.dc ? "success" : "failure"
-				rollClasses[higherIndex] = outcome
-			} else if (data.reroll?.keep === "low") {
-				const lowerIndex = data.reroll.oldRoll > data.roll! ? 1 : 0
-				const outcome = rolls[lowerIndex]! >= data.dc ? "success" : "failure"
-				rollClasses[lowerIndex] = outcome
-			} else {
-				const outcome = data.roll ? (data.roll >= data.dc ? "success" : "failure") : ""
-				rollClasses[1] = outcome
-			}
-
-			const rerollIcon = data.reroll
-				? `<span class="fc-icon"><i class="${REROLL_ICONS[data.reroll.keep]}"></i></span>`
-				: ""
-
-			return `<div class="fc-check">
-				<span class="fc-label">
-				  <span>${data.label}</span>
-					${data.description ? `<span class="fc-description">${data.description}</span>` : ""}
-				</span>
-				<span class="fc-dc">${translate("flat.message.dc", { value: data.dc })}</span>
-				<span class="fc-roll">
-					<span class="fc-rolls">
-						<span class="${rollClasses[0]}">${rolls[0] ?? ""}</span>
-						<span class="${rollClasses[1]}">${rolls[1] ?? ""}</span>
-					</span>
-					${rerollIcon}
-				</span>
-			  <button class="${buttonClass}" data-action="roll-flatcheck" data-key="${data.key}" data-dc="${data.dc}">
-					<i class="fa-solid ${buttonIcon}"></i>
-				</button>
-			</div>`
-		})
-
-		const buttonNode = jQuery.parseHTML(
-			`<section class="fc-flatcheck-buttons">${buttonHtml.join("")}</section>`,
-		)[0] as HTMLElement
+		const buttonNode = parseHTML(buttonHtml)
 
 		if (data.grabbed && data.stupefied) {
-			$(buttonNode).append(`<div class="fc-rule-note">
+			const note = parseHTML(`<div class="fc-rule-note">
 					<span data-tooltip='"${translate("flat.message.tooltip-highest-dc")}"'><i class="fa-solid fa-circle-info"></i></span>
 				</div>`)
+			buttonNode.append(note)
 		}
 		;(() => {
 			let section = html.querySelector("section.card-buttons")
@@ -220,7 +199,7 @@ async function handleFlatButtonClick(msg: ChatMessagePF2e, key: string, dc: numb
 				${heroPoints > 0 ? `<label><input type="radio" name="choice" value="hero" checked> <i class="fa-solid fa-hospital-symbol"></i> ${translate("flat.message.reroll-hero-point")}</label>` : ""}
 				<label><input type="radio" name="choice" value="new" ${heroPoints <= 0 ? "checked" : ""}> <i class="fa-solid fa-dice"></i> ${translate("flat.message.reroll-new-result")}</label>
 				<label><input type="radio" name="choice" value="low"> <i class="fa-solid fa-dice-one"></i> ${translate("flat.message.reroll-lower-result")}</label>
-				<label><input type="radio" name="choice" value="higher"> <i class="fa-solid fa-dice-six"></i> ${translate("flat.message.reroll-higher-result")}</label>
+				<label><input type="radio" name="choice" value="high"> <i class="fa-solid fa-dice-six"></i> ${translate("flat.message.reroll-higher-result")}</label>
 			`,
 			buttons: [
 				{
@@ -289,47 +268,7 @@ function shouldShowFlatChecks(msg: ChatMessagePF2e): boolean {
 export async function preCreateMessage(msg: ChatMessagePF2e) {
 	if (!msg.actor || !shouldShowFlatChecks(msg)) return
 
-	const data: ButtonsFlags = {}
-
-	const { ignored, experimental } = flatMessageConfig.toSets()
-	if (
-		!ignored.has("manipulate") &&
-		msg.actor?.conditions.stored.some((c) => c.slug === "grabbed") &&
-		msg.item?.system.traits.value?.some((t) => t === "manipulate")
-	) {
-		data.grabbed = { label: translate("flat.message.grabbed"), dc: 5 }
-	}
-
-	if (
-		!ignored.has("deafened") &&
-		msg.actor?.conditions.stored.some((c) => c.slug === "deafened") &&
-		msg.item?.system.traits.value?.some((t) => t === "auditory")
-	) {
-		data.deafened = { label: translate("flat.message.deafened"), dc: 5 }
-	}
-	if (
-		!ignored.has("deafened-spellcasting") &&
-		msg.actor?.conditions.stored.some((c) => c.slug === "deafened") &&
-		msg.flags?.pf2e?.origin?.type === "spell" &&
-		!msg.item?.system.traits.value?.some((t) => t === "subtle")
-	) {
-		data.deafened = { label: translate("flat.message.deafened"), dc: 5 }
-	}
-
-	if (!ignored.has("stupefied") && msg.flags?.pf2e?.origin?.type === "spell") {
-		const stupefied = msg.actor?.conditions.stupefied?.value
-		if (stupefied) {
-			data.stupefied = {
-				label: translate("flat.message.stupefied", { value: stupefied }),
-				dc: 5 + stupefied,
-			}
-		}
-	}
-
-	if (!ignored.has("target")) {
-		const targetCheck = msg.actor.isOfType("creature") ? flatCheckForUserTargets(msg.actor) : null
-		if (targetCheck) data.targets = targetCheck
-	}
+	const data = collectFlatChecks(msg) as MsgFlagData
 
 	msg.updateSource({
 		[`flags.${MODULE_ID}.flatchecks`]: data,