+++ src/modules/flat/rules/setup.ts
@@ -1,0 +1,5 @@
+import { buildModifyFlatDCRuleElement } from "./modify"
+
+export function setupRuleElements() {
+	game.pf2e.RuleElements.custom["fc-ModifyFlatDC"] = buildModifyFlatDCRuleElement()
+}