+++ src/modules/flat/target.ts
@@ -1,150 +1,137 @@
-import type { ActorPF2e, CreaturePF2e, TokenPF2e } from "foundry-pf2e"
-import { translate } from "src/utils"
+import type { ActorPF2e, TokenDocumentPF2e, TokenPF2e } from "foundry-pf2e"
+import * as R from "remeda"
 import { tokenLightLevel } from "./light/token"
 import { LightLevels } from "./light/utils"
 import { flatMessageConfig } from "./message-config"
-
-const originConditionDCs = {
-	dazzled: 5,
-	blinded: 11,
-}
-type OriginConditionSlug = keyof typeof originConditionDCs
 
 const originToTargetCondition = Object.freeze({
 	dazzled: "concealed",
 	blinded: "hidden",
 })
+// type OriginConditionSlug = keyof typeof originToTargetCondition
 
 export const targetConditionDCs = {
 	concealed: 5,
 	hidden: 11,
-	invisible: 11,
+	/* invisible: 11,
 	undetected: 11,
-	unnoticed: 11,
+	unnoticed: 11, */
 }
 type TargetConditionSlug = keyof typeof targetConditionDCs
 
-export function flatCheckForUserTargets(origin: CreaturePF2e) {
+/* export function flatCheckForUserTargets(origin: CreaturePF2e) {
 	if (game.user.targets.size > 1) {
 		const count = game.user.targets.reduce(
-			(acc, t) => (calculateFlatCheck(origin, t) !== null ? acc + 1 : acc),
+			(acc, t) => (calculateTargetFlatCheck(origin, t.document) !== null ? acc + 1 : acc),
 			0,
 		)
 		if (count) return { count }
 	} else if (game.user.targets.size === 1) {
 		const target = game.user.targets.first()
 		if (target) {
-			const targetDC = calculateFlatCheck(origin, target)
+			const targetDC = calculateTargetFlatCheck(origin, target.document)
 			if (targetDC) {
 				return targetDC
 			}
 		}
 	}
+	return null
+} */
+
+const sourcePriorities = [
+	"blinded",
+	"darkness",
+	"hidden",
+	"dim-light",
+	"dazzled",
+	"concealed",
+] as const
+
+interface FlatCheckSource {
+	condition: TargetConditionSlug
+	source: (typeof sourcePriorities)[number]
 }
 
-function flatCheckDataForOrigin(origin: CreaturePF2e): ConditionData | null {
-	let originCondition = null as OriginConditionSlug | null
-	origin.conditions.stored.forEach((c) => {
-		const slug = c.system.slug
-		if (
-			slug in originConditionDCs &&
-			(!originCondition || originConditionDCs[originCondition] < originConditionDCs[slug])
-		) {
-			originCondition = slug as OriginConditionSlug
-		}
+function keepHigherSource(...sources: (FlatCheckSource | null | undefined)[]) {
+	return R.firstBy(sources, (s) => {
+		if (s == null) return Infinity
+		const p = sourcePriorities.indexOf(s.source)
+		if (p === -1) return Infinity
+		return p
 	})
-
-	return originCondition
-		? {
-				slug: originToTargetCondition[originCondition],
-				dc: originConditionDCs[originCondition],
-				description: translate(`flat.target.condition.${originCondition}`),
-			}
-		: null
 }
 
-export function flatCheckDataForTarget(target: TokenPF2e): ConditionData | null {
-	let targetCondition = null as TargetConditionSlug | null
-	target.actor?.conditions?.stored.forEach((c) => {
-		const slug = c.system.slug
-		if (
-			slug in targetConditionDCs &&
-			(!targetCondition || targetConditionDCs[targetCondition] < targetConditionDCs[slug])
-		)
-			targetCondition = slug as TargetConditionSlug
-	})
-
-	return targetCondition ? { slug: targetCondition, dc: targetConditionDCs[targetCondition] } : null
-}
-
-interface ConditionData {
-	slug: OriginConditionSlug | TargetConditionSlug
-	dc: number
-	description?: string
-}
-
-interface DisplayData {
-	label: string
-	dc: number
-	description?: string
-}
-
-export function conditionToDisplayData(condition: ConditionData): DisplayData {
-	return {
-		label: translate(`flat.target.condition.${condition.slug}`),
-		dc: condition.dc,
-		description: condition.description,
+function flatCheckDataFromOrigin(origin: ActorPF2e): FlatCheckSource | null {
+	for (const slug of ["blinded", "dazzled"] as const) {
+		if (origin.conditions.bySlug(slug).length)
+			return { condition: originToTargetCondition[slug], source: slug }
 	}
+	return null
 }
 
-export function calculateFlatCheck(
-	origin: CreaturePF2e | null,
-	target: TokenPF2e,
-): DisplayData | null {
-	const perceptionActive = game.modules.get("pf2e-perception")?.active
-	// Flat check with no origin is not supported by pf2e-perception
-	if (!origin && perceptionActive) return null
-	if (origin && perceptionActive) {
-		const perceptionApi = (game.modules.get("pf2e-perception") as any).api
-		const originToken = canvas.tokens.placeables.find((t) => t.actor?.uuid === origin.uuid)
-		const condition = perceptionApi.token.getVisibility(target, originToken, {
-			affects: "target",
-		}) as TargetConditionSlug
-		const dc = perceptionApi.check.getFlatCheckDc(originToken, target) as number
-
-		if (dc === 0) return null
-		return { label: translate(`flat.target.condition.${condition}`), dc }
+export function flatCheckDataForTarget(target: ActorPF2e): FlatCheckSource | null {
+	// Todo: Invisibility
+	for (const slug of ["hidden", "concealed"] as const) {
+		if (target.conditions.bySlug(slug).length) return { condition: slug, source: slug }
 	}
+	return null
+}
 
-	const originCondition = origin ? flatCheckDataForOrigin(origin) : null
-	let targetCondition = flatCheckDataForTarget(target)
+export function visioneerFlatCheck(origin: TokenDocumentPF2e, target: TokenDocumentPF2e) {
+	const visioneerApi:
+		| { getVisibility: (observerId: string, targetId: string) => string | null }
+		| undefined =
+		// @ts-expect-error
+		game.modules.get("pf2e-visioner")?.api
 
-	if (flatMessageConfig.toSets().experimental.has("light-level")) {
-		const lightLevelCondition = conditionFromLightLevel(origin, target)
-		if (lightLevelCondition && (!targetCondition || targetCondition.dc < lightLevelCondition.dc)) {
-			targetCondition = {
-				slug: lightLevelCondition.slug,
-				dc: lightLevelCondition.dc,
-				description: lightLevelCondition.description,
-			}
+	if (visioneerApi) {
+		const condition = visioneerApi.getVisibility(origin.id, target.id)
+
+		// visioner also return undetected (and who knows what else)
+		// Make sure to only return concealed/hidden or everything explodes
+		let safeCondition: "hidden" | "concealed" | null = null
+		if (condition === "concealed" || condition === "hidden") safeCondition = condition
+		else if (condition != null) safeCondition = "hidden"
+
+		if (safeCondition) {
+			return { condition: safeCondition, source: safeCondition }
 		}
 	}
+	return null
+}
+
+export function calculateTargetFlatCheck(
+	origin: TokenDocumentPF2e | null,
+	target: TokenDocumentPF2e,
+) {
+	const originCondition = origin?.actor ? flatCheckDataFromOrigin(origin.actor) : null
+	const targetCondition = target.actor ? flatCheckDataForTarget(target.actor) : null
+
+	let source = keepHigherSource(originCondition, targetCondition)
 
 	if (
-		(targetCondition && !originCondition) ||
-		(targetCondition && originCondition && originCondition.dc < targetCondition.dc)
+		flatMessageConfig.toSets().experimental.has("light-level") &&
+		target.object &&
+		origin?.actor
 	) {
-		return conditionToDisplayData(targetCondition)
-	} else if (originCondition) {
-		return conditionToDisplayData(originCondition)
+		const lightLevelCondition = conditionFromLightLevel(origin.actor, target.object)
+		source = keepHigherSource(source, lightLevelCondition)
 	}
+
+	if (origin && target) {
+		const check = visioneerFlatCheck(origin, target)
+		if (check) source = keepHigherSource(source, check)
+	}
+
+	if (source) return { ...source, baseDc: targetConditionDCs[source.condition] }
+
 	return null
 }
 
 export function conditionFromLightLevel(
 	origin: ActorPF2e | null,
 	token: TokenPF2e,
-): { description: string; slug: TargetConditionSlug; dc: number } | null {
+): FlatCheckSource | null {
 	let hasDarkvision = false
 	let hasLowLightVision = false
 	if (origin?.isOfType("creature")) {
@@ -157,33 +144,32 @@ export function conditionFromLightLevel(
 	const lightLevel = tokenLightLevel(token)
 	if (lightLevel === LightLevels.DARK && !hasDarkvision)
 		return {
-			description: translate("flat.target.darkness"),
-			slug: "hidden",
-			dc: targetConditionDCs.hidden,
+			source: "darkness",
+			condition: "hidden",
 		}
 	if (lightLevel === LightLevels.DIM && !hasLowLightVision)
 		return {
-			description: translate("flat.target.dim-light"),
-			slug: "concealed",
-			dc: targetConditionDCs.concealed,
+			source: "dim-light",
+			condition: "concealed",
 		}
 	return null
 }
 
-export function guessOrigin(): CreaturePF2e | null {
+export function guessOrigin(): TokenDocumentPF2e | null {
 	if (canvas.tokens.controlled.length === 1) {
 		const token = canvas.tokens.controlled[0]
-		return token.actor?.isOfType("creature") ? token.actor : null
+		return token.actor?.isOfType("creature") ? token.document : null
 	} else if (canvas.tokens.controlled.length > 1) {
 		return null
 	}
 	if (game.user.isGM) return null
 
-	let actor: CreaturePF2e | null = null
+	// If there is only one token on the scene the user owns, use that one
+	let token: TokenDocumentPF2e | null = null
 	for (const t of canvas.tokens.placeables) {
 		if (!t.actor?.isOwner || !t.actor?.isOfType("creature")) continue
-		if (actor) return null
-		actor = t.actor
+		if (token) return null
+		token = t.document
 	}
-	return actor
+	return token
 }