+++ src/utils.ts
@@ -33,3 +33,20 @@ export function translate(key: string, data?: Record<string, string | number>) {
 	if (!key.startsWith("pf2e-fc.")) key = `pf2e-fc.${key}`
 	return data ? game.i18n.format(key, data) : game.i18n.localize(key)
 }
+
+type MissingOpts = {
+	prefix?: string
+	data?: Record<string, string | number>
+	case?: "title"
+}
+export function translateHandleMissing(key: string, opts: MissingOpts) {
+	const fullKey = opts.prefix ? `${opts.prefix}${key}` : key
+	const translation = opts.data ? game.i18n.format(fullKey, opts.data) : game.i18n.localize(fullKey)
+	if (fullKey !== translation) return translation
+
+	if (opts.case === "title")
+		return key.replace(/\w\S*/g, (t) => {
+			return t.charAt(0).toUpperCase() + t.substring(1).toLowerCase()
+		})
+	return key
+}